Remove RPC clients into separate module
