+++ src/client/node_client.rs
@@ -6,9 +6,9 @@ use flatbuffers::FlatBufferBuilder;
 use thread_local::CachedThreadLocal;
 
 use crate::base::utils::{decode_fast_read_response_inplace, finalize_request, response_or_error};
+use crate::client::tcp_client::TcpClient;
 use crate::generated::*;
 use crate::storage::ROOT_INODE;
-use crate::tcp_client::TcpClient;
 use fuse::FileAttr;
 use std::ops::Add;
 use std::time::{Duration, SystemTime};
