+++ src/server/fsck_handler.rs
@@ -1,6 +1,6 @@
 use crate::base::utils::{empty_response, FlatBufferResponse, ResultResponse};
+use crate::client::PeerClient;
 use crate::generated::*;
-use crate::peer_client::PeerClient;
 use crate::server::storage_node::LocalContext;
 use crate::storage::raft_group_manager::LocalRaftGroupManager;
 use crate::storage::raft_node::sync_with_leader;
