+++ src/server/router.rs
@@ -4,6 +4,7 @@ use crate::base::message_utils::{
 use crate::base::utils::{
     empty_response, finalize_response, FlatBufferResponse, FlatBufferWithResponse,
 };
+use crate::client::RemoteRaftGroups;
 use crate::generated::*;
 use crate::server::fsck_handler::{checksum_request, fsck};
 use crate::server::router::FullOrPartialResponse::{Full, Partial};
@@ -12,7 +13,7 @@ use crate::server::transaction_coordinator::{
     create_transaction, hardlink_transaction, rename_transaction, rmdir_transaction,
     unlink_transaction,
 };
-use crate::storage::raft_group_manager::{LocalRaftGroupManager, RemoteRaftGroups};
+use crate::storage::raft_group_manager::LocalRaftGroupManager;
 use crate::storage::raft_node::sync_with_leader;
 use flatbuffers::FlatBufferBuilder;
 use protobuf::Message as ProtobufMessage;
