+++ src/server/transaction_coordinator.rs
@@ -2,8 +2,9 @@ use crate::base::utils::{
     check_access, empty_response, finalize_request_without_prefix, finalize_response,
     response_or_error, FlatBufferWithResponse,
 };
+use crate::client::RemoteRaftGroups;
 use crate::generated::*;
-use crate::storage::raft_group_manager::{LocalRaftGroupManager, RemoteRaftGroups};
+use crate::storage::raft_group_manager::LocalRaftGroupManager;
 use flatbuffers::{FlatBufferBuilder, SIZE_UOFFSET};
 use rand::Rng;
 use std::collections::HashSet;
