+++ src/storage/raft_node.rs
@@ -9,8 +9,8 @@ use crate::base::message_utils::{access_type, accessed_inode, request_locks};
 use crate::base::utils::{
     empty_response, node_id_from_address, FlatBufferResponse, ResultResponse,
 };
+use crate::client::PeerClient;
 use crate::generated::*;
-use crate::peer_client::PeerClient;
 use crate::server::LocalContext;
 use crate::storage::local::FileStorage;
 use crate::storage::lock_table::LockTable;
